test(whats_new): consolidate three suites into one to cut isolate startup
